requirements.txt
Flask==2.3.2
pymongo==4.6.0
flask-jwt-extended==4.5.3
email-validator==1.3.1
python-dotenv==1.0.0